Create dac

+++ src/uart_manager.h
@@ -9,8 +9,8 @@ void print_UART(char * s);
 
 //void UART_puts(USART_TypeDef* USARTx, volatile char *s);
 
-void LOGD(char * text, uint8_t code);
-void LOGE(char * text, uint8_t code);
-void LOGI(char * text, uint8_t code);
+void LOGD(char * text, uint16_t code);
+void LOGE(char * text, uint16_t code);
+void LOGI(char * text, uint16_t code);
 
 #endif /* UARTMANAGER_H_INCLUDED */
